move plot functions to individual module
